moves eui mapping to own file (#52518)
